backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
ezdxf==1.4.2
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
orjson==3.10.15
pywin32==308; platform_system == "Windows"
google-genai>=1.65.0
Pillow>=10.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
